+++ src/PipeTransport.ts
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 import {helper, debugError, PuppeteerEventListener} from './helper';
+import type {ConnectionTransport} from './ConnectionTransport';
 
-export class PipeTransport implements Puppeteer.ConnectionTransport {
+export class PipeTransport implements ConnectionTransport {
   _pipeWrite: NodeJS.WritableStream;
   _pendingMessage: string;
   _eventListeners: PuppeteerEventListener[];
